requirements.txt
streamlit
openai
yfinance
requests
python-dotenv
